Added SmartEnemy which tracks player movement, converted all intager position values to floats.

+++ SimpleGame/src/com/djdarkside/simple/GameObject.java
@@ -5,13 +5,13 @@ import java.awt.Rectangle;
 
 public abstract class GameObject {
 
-	protected int x;
-	protected int y;
+	protected float x;
+	protected float y;
 	protected ID id;
-	protected int velX;
-	protected int velY;
+	protected float velX;
+	protected float velY;
 	
-	public GameObject(int x, int y, ID id) {
+	public GameObject(float x, float y, ID id) {
 		this.x = x;
 		this.y = y;
 		this.id = id;
@@ -27,10 +27,10 @@ public abstract class GameObject {
 	public void setY(int y) {
 		this.y = y;
 	}
-	public int getX() {
+	public float getX() {
 		return x;
 	}
-	public int getY() {
+	public float getY() {
 		return y;
 	}
 	public void setID(ID id) {
@@ -45,10 +45,10 @@ public abstract class GameObject {
 	public void setVelY(int velY) {
 		this.velY = velY;
 	}
-	public int getVelX() {
+	public float getVelX() {
 		return velX;
 	}
-	public int getVelY() {
+	public float getVelY() {
 		return velY;
 	}
